feat(refactor): block and shortcode
