+++ wp-super-duper.php
@@ -5,6 +5,8 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 if ( ! class_exists( 'WP_Super_Duper' ) ) {
 
+	require_once( 'type/shortcode.php' );
+	require_once( 'type/block.php' );
 
 	/**
 	 * A Class to be able to create a Widget, Shortcode or Block to be able to output content for WordPress.
@@ -16,8 +18,7 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 	 * @ver 1.0.19
 	 */
 	class WP_Super_Duper extends WP_Widget {
-
-		public $version = "1.0.26";
+		public $version = "1.0.24";
 		public $font_awesome_icon_version = "5.11.2";
 		public $block_code;
 		public $options;
@@ -62,15 +63,10 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 				$this->class_name = $options['class_name'];
 
 				// register shortcode
-				$this->register_shortcode();
-
-				// Fusion Builder (avada) support
-				if ( function_exists( 'fusion_builder_map' ) ) {
-					add_action( 'init', array( $this, 'register_fusion_element' ) );
-				}
+				New WP_Super_Duper_Shortcode( $this );
 
 				// register block
-				add_action( 'admin_enqueue_scripts', array( $this, 'register_block' ) );
+				New WP_Super_Duper_Block( $this );
 			}
 
 			// add the CSS and JS we need ONCE
@@ -86,25 +82,7 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 
 				$sd_widget_scripts = true;
 
-				// add shortcode insert button once
-				add_action( 'media_buttons', array( $this, 'shortcode_insert_button' ) );
-				// generatepress theme sections compatibility
-				if ( function_exists( 'generate_sections_sections_metabox' ) ) {
-					add_action( 'generate_sections_metabox', array( $this, 'shortcode_insert_button_script' ) );
-				}
-				if ( $this->is_preview() ) {
-					add_action( 'wp_footer', array( $this, 'shortcode_insert_button_script' ) );
-					// this makes the insert button work for elementor
-					add_action( 'elementor/editor/after_enqueue_scripts', array(
-						$this,
-						'shortcode_insert_button_script'
-					) ); // for elementor
-				}
-				// this makes the insert button work for cornerstone
-				add_action( 'wp_print_footer_scripts', array( __CLASS__, 'maybe_cornerstone_builder' ) );
-
 				add_action( 'wp_ajax_super_duper_get_widget_settings', array( __CLASS__, 'get_widget_settings' ) );
-				add_action( 'wp_ajax_super_duper_get_picker', array( __CLASS__, 'get_picker' ) );
 
 				// add generator text to admin head
 				add_action( 'admin_head', array( $this, 'generator' ) );
@@ -118,192 +96,6 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 		 */
 		public function elementor_editor_styles() {
 			wp_add_inline_style( 'elementor-editor', $this->widget_css( false ) );
-		}
-
-		public function register_fusion_element() {
-
-			$options = $this->options;
-
-			if ( $this->base_id ) {
-
-				$params = $this->get_fusion_params();
-
-				$args = array(
-					'name'            => $options['name'],
-					'shortcode'       => $this->base_id,
-					'icon'            => $options['block-icon'] ? $options['block-icon'] : 'far fa-square',
-					'allow_generator' => true,
-				);
-
-				if ( ! empty( $params ) ) {
-					$args['params'] = $params;
-				}
-
-				fusion_builder_map( $args );
-			}
-
-		}
-
-		public function get_fusion_params() {
-			$params    = array();
-			$arguments = $this->get_arguments();
-
-			if ( ! empty( $arguments ) ) {
-				foreach ( $arguments as $key => $val ) {
-					$param = array();
-					// type
-					$param['type'] = str_replace(
-						array(
-							"text",
-							"number",
-							"email",
-							"color",
-							"checkbox"
-						),
-						array(
-							"textfield",
-							"textfield",
-							"textfield",
-							"colorpicker",
-							"select",
-
-						),
-						$val['type'] );
-
-					// multiselect
-					if ( $val['type'] == 'multiselect' || ( ( $param['type'] == 'select' || $val['type'] == 'select' ) && ! empty( $val['multiple'] ) ) ) {
-						$param['type']     = 'multiple_select';
-						$param['multiple'] = true;
-					}
-
-					// heading
-					$param['heading'] = $val['title'];
-
-					// description
-					$param['description'] = isset( $val['desc'] ) ? $val['desc'] : '';
-
-					// param_name
-					$param['param_name'] = $key;
-
-					// Default
-					$param['default'] = isset( $val['default'] ) ? $val['default'] : '';
-
-					// Group
-					if ( isset( $val['group'] ) ) {
-						$param['group'] = $val['group'];
-					}
-
-					// value
-					if ( $val['type'] == 'checkbox' ) {
-						if ( isset( $val['default'] ) && $val['default'] == '0' ) {
-							unset( $param['default'] );
-						}
-						$param['value'] = array( '' => __( "No" ), '1' => __( "Yes" ) );
-					} elseif ( $param['type'] == 'select' || $param['type'] == 'multiple_select' ) {
-						$param['value'] = isset( $val['options'] ) ? $val['options'] : array();
-					} else {
-						$param['value'] = isset( $val['default'] ) ? $val['default'] : '';
-					}
-
-					// setup the param
-					$params[] = $param;
-
-				}
-			}
-
-
-			return $params;
-		}
-
-		/**
-		 * Maybe insert the shortcode inserter button in the footer if we are in the cornerstone builder
-		 */
-		public static function maybe_cornerstone_builder() {
-			if ( did_action( 'cornerstone_before_boot_app' ) ) {
-				self::shortcode_insert_button_script();
-			}
-		}
-
-		/**
-		 * A function to ge the shortcode builder picker html.
-		 *
-		 * @param string $editor_id
-		 *
-		 * @return string
-		 */
-		public static function get_picker( $editor_id = '' ) {
-
-			ob_start();
-			if ( isset( $_POST['editor_id'] ) ) {
-				$editor_id = esc_attr( $_POST['editor_id'] );
-			} elseif ( isset( $_REQUEST['et_fb'] ) ) {
-				$editor_id = 'main_content_content_vb_tiny_mce';
-			}
-
-			global $sd_widgets;
-			?>
-
-			<div class="sd-shortcode-left-wrap">
-				<?php
-				ksort( $sd_widgets );
-				//				print_r($sd_widgets);exit;
-				if ( ! empty( $sd_widgets ) ) {
-					echo '<select class="widefat" onchange="sd_get_shortcode_options(this);">';
-					echo "<option>" . __( 'Select shortcode' ) . "</option>";
-					foreach ( $sd_widgets as $shortcode => $class ) {
-						echo "<option value='" . esc_attr( $shortcode ) . "'>" . esc_attr( $shortcode ) . " (" . esc_attr( $class['name'] ) . ")</option>";
-					}
-					echo "</select>";
-
-				}
-				?>
-				<div class="sd-shortcode-settings"></div>
-
-			</div>
-
-			<div class="sd-shortcode-right-wrap">
-				<textarea id='sd-shortcode-output' disabled></textarea>
-				<div id='sd-shortcode-output-actions'>
-					<?php if ( $editor_id != '' ) { ?>
-						<button class="button sd-insert-shortcode-button"
-						        onclick="sd_insert_shortcode(<?php if ( ! empty( $editor_id ) ) {
-							        echo "'" . $editor_id . "'";
-						        } ?>)"><?php _e( 'Insert shortcode' ); ?></button>
-					<?php } ?>
-					<button class="button"
-					        onclick="sd_copy_to_clipboard()"><?php _e( 'Copy shortcode' ); ?></button>
-				</div>
-			</div>
-			<?php
-
-			$html = ob_get_clean();
-
-			if ( wp_doing_ajax() ) {
-				echo $html;
-				$should_die = true;
-
-				// some builder get the editor via ajax so we should not die on those ocasions
-				$dont_die = array(
-					'parent_tag',// WP Bakery
-					'avia_request' // enfold
-				);
-
-				foreach ( $dont_die as $request ) {
-					if ( isset( $_REQUEST[ $request ] ) ) {
-						$should_die = false;
-					}
-				}
-
-				if ( $should_die ) {
-					wp_die();
-				}
-
-			} else {
-				return $html;
-			}
-
-			return '';
-
 		}
 
 		/**
@@ -346,78 +138,6 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 			?>
 			<?php
 			wp_die();
-		}
-
-		/**
-		 * Insert shortcode builder button to classic editor (not inside Gutenberg, not needed).
-		 *
-		 * @since 1.0.0
-		 *
-		 * @param string $editor_id Optional. Shortcode editor id. Default null.
-		 * @param string $insert_shortcode_function Optional. Insert shotcode function. Default null.
-		 */
-		public static function shortcode_insert_button( $editor_id = '', $insert_shortcode_function = '' ) {
-			global $sd_widgets, $shortcode_insert_button_once;
-			if ( $shortcode_insert_button_once ) {
-				return;
-			}
-			add_thickbox();
-
-
-			/**
-			 * Cornerstone makes us play dirty tricks :/
-			 * All media_buttons are removed via JS unless they are two specific id's so we wrap our content in this ID so it is not removed.
-			 */
-			if ( function_exists( 'cornerstone_plugin_init' ) && ! is_admin() ) {
-				echo '<span id="insert-media-button">';
-			}
-
-			echo self::shortcode_button( 'this', 'true' );
-
-			// see opening note
-			if ( function_exists( 'cornerstone_plugin_init' ) && ! is_admin() ) {
-				echo '</span>'; // end #insert-media-button
-			}
-
-			// Add separate script for generatepress theme sections
-			if ( function_exists( 'generate_sections_sections_metabox' ) && did_action( 'generate_sections_metabox' ) ) {
-			} else {
-				self::shortcode_insert_button_script( $editor_id, $insert_shortcode_function );
-			}
-
-			$shortcode_insert_button_once = true;
-		}
-
-		/**
-		 * Gets the shortcode insert button html.
-		 *
-		 * @param string $id
-		 * @param string $search_for_id
-		 *
-		 * @return mixed
-		 */
-		public static function shortcode_button( $id = '', $search_for_id = '' ) {
-			ob_start();
-			?>
-			<span class="sd-lable-shortcode-inserter">
-				<a onclick="sd_ajax_get_picker(<?php echo $id;
-				if ( $search_for_id ) {
-					echo "," . $search_for_id;
-				} ?>);" href="#TB_inline?width=100%&height=550&inlineId=super-duper-content-ajaxed"
-				   class="thickbox button super-duper-content-open" title="Add Shortcode">
-					<span style="vertical-align: middle;line-height: 18px;font-size: 20px;"
-					      class="dashicons dashicons-screenoptions"></span>
-				</a>
-				<div id="super-duper-content-ajaxed" style="display:none;">
-					<span>Loading</span>
-				</div>
-			</span>
-
-			<?php
-			$html = ob_get_clean();
-
-			// remove line breaks so we can use it in js
-			return preg_replace( "/\r|\n/", "", trim( $html ) );
 		}
 
 		/**
@@ -516,497 +236,6 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 				'<script>',
 				'</script>'
 			), '', $output );
-		}
-
-		/**
-		 * Output the JS and CSS for the shortcode insert button.
-		 *
-		 * @since 1.0.6
-		 *
-		 * @param string $editor_id
-		 * @param string $insert_shortcode_function
-		 */
-		public static function shortcode_insert_button_script( $editor_id = '', $insert_shortcode_function = '' ) {
-			?>
-			<style>
-				.sd-shortcode-left-wrap {
-					float: left;
-					width: 60%;
-				}
-
-				.sd-shortcode-left-wrap .gd-help-tip {
-					float: none;
-				}
-
-				.sd-shortcode-left-wrap .widefat {
-					border-spacing: 0;
-					width: 100%;
-					clear: both;
-					margin: 0;
-					border: 1px solid #ddd;
-					box-shadow: inset 0 1px 2px rgba(0, 0, 0, .07);
-					background-color: #fff;
-					color: #32373c;
-					outline: 0;
-					transition: 50ms border-color ease-in-out;
-					padding: 3px 5px;
-				}
-
-				.sd-shortcode-left-wrap input[type=checkbox].widefat {
-					border: 1px solid #b4b9be;
-					background: #fff;
-					color: #555;
-					clear: none;
-					cursor: pointer;
-					display: inline-block;
-					line-height: 0;
-					height: 16px;
-					margin: -4px 4px 0 0;
-					margin-top: 0;
-					outline: 0;
-					padding: 0 !important;
-					text-align: center;
-					vertical-align: middle;
-					width: 16px;
-					min-width: 16px;
-					-webkit-appearance: none;
-					box-shadow: inset 0 1px 2px rgba(0, 0, 0, .1);
-					transition: .05s border-color ease-in-out;
-				}
-
-				.sd-shortcode-left-wrap input[type=checkbox]:checked:before {
-					content: "\f147";
-					margin: -3px 0 0 -4px;
-					color: #1e8cbe;
-					float: left;
-					display: inline-block;
-					vertical-align: middle;
-					width: 16px;
-					font: normal 21px/1 dashicons;
-					speak: none;
-					-webkit-font-smoothing: antialiased;
-					-moz-osx-font-smoothing: grayscale;
-				}
-
-				#sd-shortcode-output-actions button,
-				.sd-advanced-button {
-					color: #555;
-					border-color: #ccc;
-					background: #f7f7f7;
-					box-shadow: 0 1px 0 #ccc;
-					vertical-align: top;
-					display: inline-block;
-					text-decoration: none;
-					font-size: 13px;
-					line-height: 26px;
-					height: 28px;
-					margin: 0;
-					padding: 0 10px 1px;
-					cursor: pointer;
-					border-width: 1px;
-					border-style: solid;
-					-webkit-appearance: none;
-					border-radius: 3px;
-					white-space: nowrap;
-					box-sizing: border-box;
-				}
-
-				button.sd-advanced-button {
-					background: #0073aa;
-					border-color: #006799;
-					box-shadow: inset 0 2px 0 #006799;
-					vertical-align: top;
-					color: #fff;
-					text-decoration: none;
-					text-shadow: 0 -1px 1px #006799, 1px 0 1px #006799, 0 1px 1px #006799, -1px 0 1px #006799;
-					float: right;
-					margin-right: 3px !important;
-					font-size: 20px !important;
-				}
-
-				.sd-shortcode-right-wrap {
-					float: right;
-					width: 35%;
-				}
-
-				#sd-shortcode-output {
-					background: rgba(255, 255, 255, .5);
-					border-color: rgba(222, 222, 222, .75);
-					box-shadow: inset 0 1px 2px rgba(0, 0, 0, .04);
-					color: rgba(51, 51, 51, .5);
-					overflow: auto;
-					padding: 2px 6px;
-					line-height: 1.4;
-					resize: vertical;
-				}
-
-				#sd-shortcode-output {
-					height: 250px;
-					width: 100%;
-				}
-
-				<?php if ( function_exists( 'generate_sections_sections_metabox' ) ) { ?>
-				.generate-sections-modal #custom-media-buttons > .sd-lable-shortcode-inserter {
-					display: inline;
-				}
-
-				<?php } ?>
-			</style>
-			<?php
-			if ( class_exists( 'SiteOrigin_Panels' ) ) {
-				echo "<script>" . self::siteorigin_js() . "</script>";
-			}
-			?>
-			<script>
-				<?php
-				if(! empty( $insert_shortcode_function )){
-					echo $insert_shortcode_function;
-				}else{
-
-				/**
-				 * Function for super duper insert shortcode.
-				 *
-				 * @since 1.0.0
-				 */
-				?>
-				function sd_insert_shortcode($editor_id) {
-					$shortcode = jQuery('#TB_ajaxContent #sd-shortcode-output').val();
-					if ($shortcode) {
-						if (!$editor_id) {
-							<?php
-							if ( isset( $_REQUEST['et_fb'] ) ) {
-								echo '$editor_id = "#main_content_content_vb_tiny_mce";';
-							} elseif ( isset( $_REQUEST['action'] ) && $_REQUEST['action'] == 'elementor' ) {
-								echo '$editor_id = "#elementor-controls .wp-editor-container textarea";';
-							} else {
-								echo '$editor_id = "#wp-content-editor-container textarea";';
-							}
-							?>
-						} else {
-							$editor_id = '#' + $editor_id;
-						}
-						tmceActive = jQuery($editor_id).attr("aria-hidden") == "true" ? true : false;
-						/* GeneratePress */
-						if (jQuery('#generate-sections-modal-dialog ' + $editor_id).length) {
-							$editor_id = '#generate-sections-modal-dialog ' + $editor_id;
-							tmceActive = jQuery($editor_id).closest('.wp-editor-wrap').hasClass('tmce-active') ? true : false;
-						}
-						if (tinyMCE && tinyMCE.activeEditor && tmceActive) {
-							tinyMCE.execCommand('mceInsertContent', false, $shortcode);
-						} else {
-							var $txt = jQuery($editor_id);
-							var caretPos = $txt[0].selectionStart;
-							var textAreaTxt = $txt.val();
-							var txtToAdd = $shortcode;
-							var textareaValue = textAreaTxt.substring(0, caretPos) + txtToAdd + textAreaTxt.substring(caretPos);
-							$txt.focus().val(textareaValue).change().keydown().blur().keyup().keypress().trigger('input').trigger('change');
-
-							// set Divi react input value
-							var input = document.getElementById("main_content_content_vb_tiny_mce");
-							if (input) {
-								sd_setNativeValue(input, textareaValue);
-							}
-
-						}
-						tb_remove();
-					}
-				}
-
-				/*
-				 Set the value of elements controled via react.
-				 */
-				function sd_setNativeValue(element, value) {
-					let lastValue = element.value;
-					element.value = value;
-					let event = new Event("input", {target: element, bubbles: true});
-					// React 15
-					event.simulated = true;
-					// React 16
-					let tracker = element._valueTracker;
-					if (tracker) {
-						tracker.setValue(lastValue);
-					}
-					element.dispatchEvent(event);
-				}
-				<?php }?>
-
-				/*
-				 Copies the shortcode to the clipboard.
-				 */
-				function sd_copy_to_clipboard() {
-					/* Get the text field */
-					var copyText = document.querySelector("#TB_ajaxContent #sd-shortcode-output");
-					//un-disable the field
-					copyText.disabled = false;
-					/* Select the text field */
-					copyText.select();
-					/* Copy the text inside the text field */
-					document.execCommand("Copy");
-					//re-disable the field
-					copyText.disabled = true;
-					/* Alert the copied text */
-					alert("Copied the text: " + copyText.value);
-				}
-
-				/*
-				 Gets the shortcode options.
-				 */
-				function sd_get_shortcode_options($this) {
-
-					$short_code = jQuery($this).val();
-					if ($short_code) {
-
-						var data = {
-							'action': 'super_duper_get_widget_settings',
-							'shortcode': $short_code,
-							'attributes': 123,
-							'post_id': 321,
-							'_ajax_nonce': '<?php echo wp_create_nonce( 'super_duper_output_shortcode' );?>'
-						};
-
-						if (typeof ajaxurl === 'undefined') {
-							var ajaxurl = "<?php echo admin_url( 'admin-ajax.php' );?>";
-						}
-
-						jQuery.post(ajaxurl, data, function (response) {
-							jQuery('#TB_ajaxContent .sd-shortcode-settings').html(response);
-
-							jQuery('#' + $short_code).on('change', 'select', function () {
-								sd_build_shortcode($short_code);
-							}); // take care of select tags
-
-							jQuery('#' + $short_code).on('change keypress keyup', 'input,textarea', function () {
-								sd_build_shortcode($short_code);
-							});
-
-							sd_build_shortcode($short_code);
-
-							// resize the window to fit
-							setTimeout(function () {
-								jQuery('#TB_ajaxContent').css('width', 'auto').css('height', '75vh');
-							}, 200);
-
-
-							return response;
-						});
-					}
-
-				}
-
-				/*
-				 Builds and inserts the shortcode into the viewer.
-				 */
-				function sd_build_shortcode($id) {
-
-					var multiSelects = {};
-					var multiSelectsRemove = [];
-
-					$output = "[" + $id;
-
-					$form_data = jQuery("#" + $id).serializeArray();
-
-					// run checks for multiselects
-					jQuery.each($form_data, function (index, element) {
-						if (element && element.value) {
-							$field_name = element.name.substr(element.name.indexOf("][") + 2);
-							$field_name = $field_name.replace("]", "");
-							// check if its a multiple
-							if ($field_name.includes("[]")) {
-								multiSelectsRemove[multiSelectsRemove.length] = index;
-								$field_name = $field_name.replace("[]", "");
-								if ($field_name in multiSelects) {
-									multiSelects[$field_name] = multiSelects[$field_name] + "," + element.value;
-								} else {
-									multiSelects[$field_name] = element.value;
-								}
-							}
-						}
-					});
-
-					// fix multiselects if any are found
-					if (multiSelectsRemove.length) {
-
-						// remove all multiselects
-						multiSelectsRemove.reverse();
-						multiSelectsRemove.forEach(function (index) {
-							$form_data.splice(index, 1);
-						});
-
-						$ms_arr = [];
-						// add multiselets back
-						jQuery.each(multiSelects, function (index, value) {
-							$ms_arr[$ms_arr.length] = {"name": "[][" + index + "]", "value": value};
-						});
-						$form_data = $form_data.concat($ms_arr);
-					}
-
-
-					if ($form_data) {
-						$content = '';
-						$form_data.forEach(function (element) {
-
-							if (element.value) {
-								$field_name = element.name.substr(element.name.indexOf("][") + 2);
-								$field_name = $field_name.replace("]", "");
-								if ($field_name == 'html') {
-									$content = element.value;
-								} else {
-									$output = $output + " " + $field_name + '="' + element.value + '"';
-								}
-							}
-
-						});
-					}
-					$output = $output + "]";
-
-					// check for content field
-					if ($content) {
-						$output = $output + $content + "[/" + $id + "]";
-					}
-
-					jQuery('#TB_ajaxContent #sd-shortcode-output').html($output);
-				}
-
-
-				/*
-				 Delay the init of the textareas for 1 second.
-				 */
-				(function () {
-					setTimeout(function () {
-						sd_init_textareas();
-					}, 1000);
-				})();
-
-				/*
-				 Init the textareas to be able to show the shortcode builder button.
-				 */
-				function sd_init_textareas() {
-
-					// General textareas
-					jQuery(document).on('focus', 'textarea', function () {
-
-						if (jQuery(this).hasClass('wp-editor-area')) {
-							// insert the shortcode button to the textarea lable if not there already
-							if (!jQuery(this).parent().find('.sd-lable-shortcode-inserter').length) {
-								jQuery(this).parent().find('.quicktags-toolbar').append(sd_shortcode_button(jQuery(this).attr('id')));
-							}
-						} else {
-							// insert the shortcode button to the textarea lable if not there already
-							if (!jQuery("label[for='" + jQuery(this).attr('id') + "']").find('.sd-lable-shortcode-inserter').length) {
-								jQuery("label[for='" + jQuery(this).attr('id') + "']").append(sd_shortcode_button(jQuery(this).attr('id')));
-							}
-						}
-					});
-
-					// The below tries to add the shorcode builder button to the builders own raw/shortcode sections.
-
-					// DIVI
-					jQuery(document).on('focusin', '.et-fb-codemirror', function () {
-						// insert the shortcode button to the textarea lable if not there already
-						if (!jQuery(this).closest('.et-fb-form__group').find('.sd-lable-shortcode-inserter').length) {
-							jQuery(this).closest('.et-fb-form__group').find('.et-fb-form__label-text').append(sd_shortcode_button());
-						}
-					});
-
-					// Beaver
-					jQuery(document).on('focusin', '.fl-code-field', function () {
-						// insert the shortcode button to the textarea lable if not there already
-						if (!jQuery(this).closest('.fl-field-control-wrapper').find('.sd-lable-shortcode-inserter').length) {
-							jQuery(this).closest('.fl-field-control-wrapper').prepend(sd_shortcode_button());
-						}
-					});
-
-					// Fushion builder (avada)
-					jQuery(document).on('focusin', '.CodeMirror.cm-s-default', function () {
-						// insert the shortcode button to the textarea lable if not there already
-						if (!jQuery(this).parent().find('.sd-lable-shortcode-inserter').length) {
-							jQuery(sd_shortcode_button()).insertBefore(this);
-						}
-					});
-
-					// Avia builder (enfold)
-					jQuery(document).on('focusin', '#aviaTBcontent', function () {
-						// insert the shortcode button to the textarea lable if not there already
-						if (!jQuery(this).parent().parent().find('.avia-name-description ').find('.sd-lable-shortcode-inserter').length) {
-							jQuery(this).parent().parent().find('.avia-name-description strong').append(sd_shortcode_button(jQuery(this).attr('id')));
-						}
-					});
-
-					// Cornerstone textareas
-					jQuery(document).on('focusin', '.cs-control.cs-control-textarea', function () {
-						// insert the shortcode button to the textarea lable if not there already
-						if (!jQuery(this).find('.cs-control-header label').find('.sd-lable-shortcode-inserter').length) {
-							jQuery(this).find('.cs-control-header label').append(sd_shortcode_button());
-						}
-					});
-
-					// Cornerstone main bar
-					setTimeout(function () {
-						// insert the shortcode button to the textarea lable if not there already
-						if (!jQuery('.cs-bar-btns').find('.sd-lable-shortcode-inserter').length) {
-							jQuery('<li style="text-align: center;padding: 5px;list-style: none;">' + sd_shortcode_button() + '</li>').insertBefore('.cs-action-toggle-custom-css');
-						}
-					}, 2000);
-
-
-					// WP Bakery, code editor does not render shortcodes.
-//					jQuery(document).on('focusin', '.wpb-textarea_raw_html', function () {
-//						// insert the shortcode button to the textarea lable if not there already
-//						if(!jQuery(this).parent().parent().find('.wpb_element_label').find('.sd-lable-shortcode-inserter').length){
-//							jQuery(this).parent().parent().find('.wpb_element_label').append(sd_shortcode_button());
-//						}
-//					});
-
-				}
-
-				/**
-				 * Gets the html for the picker via ajax and updates it on the fly.
-				 *
-				 * @param $id
-				 * @param $search
-				 */
-				function sd_ajax_get_picker($id, $search) {
-					if ($search) {
-						$this = $id;
-						$id = jQuery($this).closest('.wp-editor-wrap').find('.wp-editor-container textarea').attr('id');
-					}
-
-					var data = {
-						'action': 'super_duper_get_picker',
-						'editor_id': $id,
-						'_ajax_nonce': '<?php echo wp_create_nonce( 'super_duper_picker' );?>'
-					};
-
-					if (!ajaxurl) {
-						var ajaxurl = "<?php echo admin_url( 'admin-ajax.php' ); ?>";
-					}
-
-					jQuery.post(ajaxurl, data, function (response) {
-						jQuery('#TB_ajaxContent').html(response);
-						//return response;
-					}).then(function (env) {
-						jQuery('body').on('thickbox:removed', function () {
-							jQuery('#super-duper-content-ajaxed').html('');
-						});
-					});
-				}
-
-				/**
-				 * Get the html for the shortcode inserter button depending on if a textarea id is available.
-				 *
-				 * @param $id string The textarea id.
-				 * @returns {string}
-				 */
-				function sd_shortcode_button($id) {
-					if ($id) {
-						return '<?php echo self::shortcode_button( "\\''+\$id+'\\'" );?>';
-					} else {
-						return '<?php echo self::shortcode_button();?>';
-					}
-				}
-
-			</script>
-			<?php
 		}
 
 		/**
@@ -1268,128 +497,6 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 		}
 
 		/**
-		 * Register the parent shortcode.
-		 *
-		 * @since 1.0.0
-		 */
-		public function register_shortcode() {
-			add_shortcode( $this->base_id, array( $this, 'shortcode_output' ) );
-			add_action( 'wp_ajax_super_duper_output_shortcode', array( __CLASS__, 'render_shortcode' ) );
-		}
-
-		/**
-		 * Render the shortcode via ajax so we can return it to Gutenberg.
-		 *
-		 * @since 1.0.0
-		 */
-		public static function render_shortcode() {
-
-			check_ajax_referer( 'super_duper_output_shortcode', '_ajax_nonce', true );
-			if ( ! current_user_can( 'manage_options' ) ) {
-				wp_die();
-			}
-
-			// we might need the $post value here so lets set it.
-			if ( isset( $_POST['post_id'] ) && $_POST['post_id'] ) {
-				$post_obj = get_post( absint( $_POST['post_id'] ) );
-				if ( ! empty( $post_obj ) && empty( $post ) ) {
-					global $post;
-					$post = $post_obj;
-				}
-			}
-
-			if ( isset( $_POST['shortcode'] ) && $_POST['shortcode'] ) {
-				$shortcode_name   = sanitize_title_with_dashes( $_POST['shortcode'] );
-				$attributes_array = isset( $_POST['attributes'] ) && $_POST['attributes'] ? $_POST['attributes'] : array();
-				$attributes       = '';
-				if ( ! empty( $attributes_array ) ) {
-					foreach ( $attributes_array as $key => $value ) {
-						if ( is_array( $value ) ) {
-							$value = implode( ",", $value );
-						}
-						$attributes .= " " . sanitize_title_with_dashes( $key ) . "='" . wp_slash( $value ) . "' ";
-					}
-				}
-
-				$shortcode = "[" . $shortcode_name . " " . $attributes . "]";
-
-				echo do_shortcode( $shortcode );
-
-			}
-			wp_die();
-		}
-
-		/**
-		 * Output the shortcode.
-		 *
-		 * @param array $args
-		 * @param string $content
-		 *
-		 * @return string
-		 */
-		public function shortcode_output( $args = array(), $content = '' ) {
-			$args = $this->argument_values( $args );
-
-			// add extra argument so we know its a output to gutenberg
-			//$args
-			$args = $this->string_to_bool( $args );
-
-			// if we have a enclosed shortcode we add it to the special `html` argument
-			if ( ! empty( $content ) ) {
-				$args['html'] = $content;
-			}
-
-			$class = isset( $this->options['widget_ops']['classname'] ) ? esc_attr( $this->options['widget_ops']['classname'] ) : '';
-			$class .= " sdel-".$this->get_instance_hash();
-
-			$class = apply_filters( 'wp_super_duper_div_classname', $class, $args, $this );
-			$class = apply_filters( 'wp_super_duper_div_classname_' . $this->base_id, $class, $args, $this );
-
-			$attrs = apply_filters( 'wp_super_duper_div_attrs', '', $args, $this );
-			$attrs = apply_filters( 'wp_super_duper_div_attrs_' . $this->base_id, '', $args, $this ); //@todo this does not seem right @kiran?
-
-			$shortcode_args = array();
-			$output         = '';
-			$no_wrap        = isset( $this->options['no_wrap'] ) && $this->options['no_wrap'] ? true : false;
-			if ( isset( $args['no_wrap'] ) && $args['no_wrap'] ) {
-				$no_wrap = true;
-			}
-			$main_content = $this->output( $args, $shortcode_args, $content );
-			if ( $main_content && ! $no_wrap ) {
-				// wrap the shortcode in a div with the same class as the widget
-				$output .= '<div class="' . $class . '" ' . $attrs . '>';
-				if ( ! empty( $args['title'] ) ) {
-					// if its a shortcode and there is a title try to grab the title wrappers
-					$shortcode_args = array( 'before_title' => '', 'after_title' => '' );
-					if ( empty( $instance ) ) {
-						global $wp_registered_sidebars;
-						if ( ! empty( $wp_registered_sidebars ) ) {
-							foreach ( $wp_registered_sidebars as $sidebar ) {
-								if ( ! empty( $sidebar['before_title'] ) ) {
-									$shortcode_args['before_title'] = $sidebar['before_title'];
-									$shortcode_args['after_title']  = $sidebar['after_title'];
-									break;
-								}
-							}
-						}
-					}
-					$output .= $this->output_title( $shortcode_args, $args );
-				}
-				$output .= $main_content;
-				$output .= '</div>';
-			} elseif ( $main_content && $no_wrap ) {
-				$output .= $main_content;
-			}
-
-			// if preview show a placeholder if empty
-			if ( $this->is_preview() && $output == '' ) {
-				$output = $this->preview_placeholder_text( "{{" . $this->base_id . "}}" );
-			}
-
-			return apply_filters( 'wp_super_duper_widget_output', $output, $args, $shortcode_args, $this );
-		}
-
-		/**
 		 * Placeholder text to show if output is empty and we are on a preview/builder page.
 		 *
 		 * @param string $name
@@ -1497,48 +604,11 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 		}
 
 		/**
-		 * Add the dynamic block code inline when the wp-block in enqueued.
-		 */
-		public function register_block() {
-			wp_add_inline_script( 'wp-blocks', $this->block() );
-			if ( class_exists( 'SiteOrigin_Panels' ) ) {
-				wp_add_inline_script( 'wp-blocks', $this->siteorigin_js() );
-			}
-		}
-
-		/**
-		 * Check if we need to show advanced options.
-		 *
-		 * @return bool
-		 */
-		public function block_show_advanced() {
-
-			$show      = false;
-			$arguments = $this->arguments;
-
-			if ( empty( $arguments ) ) {
-				$arguments = $this->get_arguments();
-			}
-
-			if ( ! empty( $arguments ) ) {
-				foreach ( $arguments as $argument ) {
-					if ( isset( $argument['advanced'] ) && $argument['advanced'] ) {
-						$show = true;
-						break; // no need to continue if we know we have it
-					}
-				}
-			}
-
-			return $show;
-		}
-
-		/**
 		 * Get the url path to the current folder.
 		 *
 		 * @return string
 		 */
 		public function get_url() {
-
 			$url = $this->url;
 
 			if ( ! $url ) {
@@ -1554,43 +624,7 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 				}
 			}
 
-
 			return $url;
-		}
-
-		/**
-		 * Generate the block icon.
-		 *
-		 * Enables the use of Font Awesome icons.
-		 *
-		 * @note xlink:href is actually deprecated but href is not supported by all so we use both.
-		 *
-		 * @param $icon
-		 *
-		 * @since 1.1.0
-		 * @return string
-		 */
-		public function get_block_icon( $icon ) {
-
-			// check if we have a Font Awesome icon
-			$fa_type = '';
-			if ( substr( $icon, 0, 7 ) === "fas fa-" ) {
-				$fa_type = 'solid';
-			} elseif ( substr( $icon, 0, 7 ) === "far fa-" ) {
-				$fa_type = 'regular';
-			} elseif ( substr( $icon, 0, 7 ) === "fab fa-" ) {
-				$fa_type = 'brands';
-			} else {
-				$icon = "'" . $icon . "'";
-			}
-
-			// set the icon if we found one
-			if ( $fa_type ) {
-				$fa_icon = str_replace( array( "fas fa-", "far fa-", "fab fa-" ), "", $icon );
-				$icon    = "el('svg',{width: 20, height: 20, viewBox: '0 0 20 20'},el('use', {'xlink:href': '" . $this->get_url() . "icons/" . $fa_type . ".svg#" . $fa_icon . "','href': '" . $this->get_url() . "icons/" . $fa_type . ".svg#" . $fa_icon . "'}))";
-			}
-
-			return $icon;
 		}
 
 		public function group_arguments( $arguments ) {
@@ -1618,729 +652,6 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 			return $arguments;
 		}
 
-
-		/**
-		 * Output the JS for building the dynamic Guntenberg block.
-		 *
-		 * @since 1.0.4 Added block_wrap property which will set the block wrapping output element ie: div, span, p or empty for no wrap.
-		 * @since 1.0.9 Save numbers as numbers and not strings.
-		 * @since 1.1.0 Font Awesome classes can be used for icons.
-		 * @return mixed
-		 */
-		public function block() {
-			ob_start();
-
-			$show_advanced = $this->block_show_advanced();
-			?>
-			<script>
-				/**
-				 * BLOCK: Basic
-				 *
-				 * Registering a basic block with Gutenberg.
-				 * Simple block, renders and saves the same content without any interactivity.
-				 *
-				 * Styles:
-				 *        editor.css — Editor styles for the block.
-				 *        style.css  — Editor & Front end styles for the block.
-				 */
-				(function () {
-					var __ = wp.i18n.__; // The __() for internationalization.
-					var el = wp.element.createElement; // The wp.element.createElement() function to create elements.
-					var editable = wp.blocks.Editable;
-					var blocks = wp.blocks;
-					var registerBlockType = wp.blocks.registerBlockType; // The registerBlockType() to register blocks.
-					var is_fetching = false;
-					var prev_attributes = [];
-
-					var term_query_type = '';
-					var post_type_rest_slugs = <?php if(! empty( $this->arguments ) && isset($this->arguments['post_type']['onchange_rest']['values'])){echo "[".json_encode($this->arguments['post_type']['onchange_rest']['values'])."]";}else{echo "[]";} ?>;
-					const taxonomies_<?php echo str_replace("-","_", $this->id);?> = [{label: "Please wait", value: 0}];
-					const sort_by_<?php echo str_replace("-","_", $this->id);?> = [{label: "Please wait", value: 0}];
-
-					/**
-					 * Register Basic Block.
-					 *
-					 * Registers a new block provided a unique name and an object defining its
-					 * behavior. Once registered, the block is made available as an option to any
-					 * editor interface where blocks are implemented.
-					 *
-					 * @param  {string}   name     Block name.
-					 * @param  {Object}   settings Block settings.
-					 * @return {?WPBlock}          The block, if it has been successfully
-					 *                             registered; otherwise `undefined`.
-					 */
-					registerBlockType('<?php echo str_replace( "_", "-", sanitize_title_with_dashes( $this->options['textdomain'] ) . '/' . sanitize_title_with_dashes( $this->options['class_name'] ) );  ?>', { // Block name. Block names must be string that contains a namespace prefix. Example: my-plugin/my-custom-block.
-						title: '<?php echo addslashes( $this->options['name'] ); ?>', // Block title.
-						description: '<?php echo addslashes( $this->options['widget_ops']['description'] )?>', // Block title.
-						icon: <?php echo $this->get_block_icon( $this->options['block-icon'] );?>,//'<?php echo isset( $this->options['block-icon'] ) ? esc_attr( $this->options['block-icon'] ) : 'shield-alt';?>', // Block icon from Dashicons → https://developer.wordpress.org/resource/dashicons/.
-						supports: {
-							<?php
-							if ( isset( $this->options['block-supports'] ) ) {
-								echo $this->array_to_attributes( $this->options['block-supports'] );
-							}
-							?>
-						},
-						category: '<?php echo isset( $this->options['block-category'] ) ? esc_attr( $this->options['block-category'] ) : 'common';?>', // Block category — Group blocks together based on common traits E.g. common, formatting, layout widgets, embed.
-						<?php if ( isset( $this->options['block-keywords'] ) ) {
-						echo "keywords : " . $this->options['block-keywords'] . ",";
-					}?>
-
-						<?php
-
-						// maybe set no_wrap
-						$no_wrap = isset( $this->options['no_wrap'] ) && $this->options['no_wrap'] ? true : false;
-						if ( isset( $this->arguments['no_wrap'] ) && $this->arguments['no_wrap'] ) {
-							$no_wrap = true;
-						}
-						if ( $no_wrap ) {
-							$this->options['block-wrap'] = '';
-						}
-
-
-
-						$show_alignment = false;
-						// align feature
-						/*echo "supports: {";
-						echo "	align: true,";
-						echo "  html: false";
-						echo "},";*/
-
-						if ( ! empty( $this->arguments ) ) {
-							echo "attributes : {";
-
-							if ( $show_advanced ) {
-								echo "show_advanced: {";
-								echo "	type: 'boolean',";
-								echo "  default: false,";
-								echo "},";
-							}
-
-							// block wrap element
-							if ( ! empty( $this->options['block-wrap'] ) ) { //@todo we should validate this?
-								echo "block_wrap: {";
-								echo "	type: 'string',";
-								echo "  default: '" . esc_attr( $this->options['block-wrap'] ) . "',";
-								echo "},";
-							}
-
-							foreach ( $this->arguments as $key => $args ) {
-
-								// set if we should show alignment
-								if ( $key == 'alignment' ) {
-									$show_alignment = true;
-								}
-
-								$extra = '';
-
-								if ( $args['type'] == 'checkbox' ) {
-									$type    = 'boolean';
-									$default = isset( $args['default'] ) && $args['default'] ? 'true' : 'false';
-								} elseif ( $args['type'] == 'number' ) {
-									$type    = 'number';
-									$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
-								} elseif ( $args['type'] == 'select' && ! empty( $args['multiple'] ) ) {
-									$type = 'array';
-									if ( isset( $args['default'] ) && is_array( $args['default'] ) ) {
-										$default = ! empty( $args['default'] ) ? "['" . implode( "','", $args['default'] ) . "']" : "[]";
-									} else {
-										$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
-									}
-								} elseif ( $args['type'] == 'multiselect' ) {
-									$type    = 'array';
-									$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
-								} else {
-									$type    = 'string';
-									$default = isset( $args['default'] ) ? "'" . $args['default'] . "'" : "''";
-								}
-								echo $key . " : {";
-								echo "type : '$type',";
-								echo "default : $default,";
-								echo "},";
-							}
-
-							echo "content : {type : 'string',default: 'Please select the attributes in the block settings'},";
-							echo "className: { type: 'string', default: '' },";
-
-							echo "},";
-
-						}
-
-						?>
-
-						// The "edit" property must be a valid function.
-						edit: function (props) {
-
-
-							var $value = '';
-							<?php
-							// if we have a post_type and a category then link them
-							if( isset($this->arguments['post_type']) && isset($this->arguments['category']) && !empty($this->arguments['category']['post_type_linked']) ){
-							?>
-							if(typeof(prev_attributes[props.id]) != 'undefined' ){
-								$pt = props.attributes.post_type;
-								if(post_type_rest_slugs.length){
-									$value = post_type_rest_slugs[0][$pt];
-								}
-								var run = false;
-
-								if($pt != term_query_type){
-									run = true;
-									term_query_type = $pt;
-								}
-
-								// taxonomies
-								if( $value && 'post_type' in prev_attributes[props.id] && 'category' in prev_attributes[props.id] && run ){
-									wp.apiFetch({path: "<?php if(isset($this->arguments['post_type']['onchange_rest']['path'])){echo $this->arguments['post_type']['onchange_rest']['path'];}else{'/wp/v2/"+$value+"/categories';} ?>"}).then(terms => {
-										while (taxonomies_<?php echo str_replace("-","_", $this->id);?>.length) {
-										taxonomies_<?php echo str_replace("-","_", $this->id);?>.pop();
-									}
-									taxonomies_<?php echo str_replace("-","_", $this->id);?>.push({label: "All", value: 0});
-									jQuery.each( terms, function( key, val ) {
-										taxonomies_<?php echo str_replace("-","_", $this->id);?>.push({label: val.name, value: val.id});
-									});
-
-									// setting the value back and fourth fixes the no update issue that sometimes happens where it won't update the options.
-									var $old_cat_value = props.attributes.category
-									props.setAttributes({category: [0] });
-									props.setAttributes({category: $old_cat_value });
-
-									return taxonomies_<?php echo str_replace("-","_", $this->id);?>;
-								});
-								}
-
-								// sort_by
-								if( $value && 'post_type' in prev_attributes[props.id] && 'sort_by' in prev_attributes[props.id] && run ){
-									var data = {
-										'action': 'geodir_get_sort_options',
-										'post_type': $pt
-									};
-									jQuery.post(ajaxurl, data, function(response) {
-										response = JSON.parse(response);
-										while (sort_by_<?php echo str_replace("-","_", $this->id);?>.length) {
-											sort_by_<?php echo str_replace("-","_", $this->id);?>.pop();
-										}
-
-										jQuery.each( response, function( key, val ) {
-											sort_by_<?php echo str_replace("-","_", $this->id);?>.push({label: val, value: key});
-										});
-
-										// setting the value back and fourth fixes the no update issue that sometimes happens where it won't update the options.
-										var $old_sort_by_value = props.attributes.sort_by
-										props.setAttributes({sort_by: [0] });
-										props.setAttributes({sort_by: $old_sort_by_value });
-
-										return sort_by_<?php echo str_replace("-","_", $this->id);?>;
-									});
-
-								}
-							}
-							<?php }?>
-
-
-							var content = props.attributes.content;
-
-							function onChangeContent() {
-
-								$refresh = false;
-
-								// Set the old content the same as the new one so we only compare all other attributes
-								if(typeof(prev_attributes[props.id]) != 'undefined'){
-									prev_attributes[props.id].content = props.attributes.content;
-								}else if(props.attributes.content === ""){
-									// if first load and content empty then refresh
-									$refresh = true;
-								}
-
-								if ( ( !is_fetching &&  JSON.stringify(prev_attributes[props.id]) != JSON.stringify(props.attributes) ) || $refresh  ) {
-
-									is_fetching = true;
-									var data = {
-										'action': 'super_duper_output_shortcode',
-										'shortcode': '<?php echo $this->options['base_id'];?>',
-										'attributes': props.attributes,
-										'post_id': <?php global $post; if ( isset( $post->ID ) ) {
-										echo $post->ID;
-									}else{echo '0';}?>,
-										'_ajax_nonce': '<?php echo wp_create_nonce( 'super_duper_output_shortcode' );?>'
-									};
-
-									jQuery.post(ajaxurl, data, function (response) {
-										return response;
-									}).then(function (env) {
-
-										// if the content is empty then we place some placeholder text
-										if (env == '') {
-											env = "<div style='background:#0185ba33;padding: 10px;border: 4px #ccc dashed;'>" + "<?php _e( 'Placeholder for: ' );?>" + props.name + "</div>";
-										}
-
-										props.setAttributes({content: env});
-										is_fetching = false;
-										prev_attributes[props.id] = props.attributes;
-
-										// if AUI is active call the js init function
-										if (typeof aui_init === "function") {
-											aui_init();
-										}
-									});
-
-
-								}
-
-								return props.attributes.content;
-
-							}
-
-							return [
-
-								el(wp.blockEditor.BlockControls, {key: 'controls'},
-
-									<?php if($show_alignment){?>
-									el(
-										wp.blockEditor.AlignmentToolbar,
-										{
-											value: props.attributes.alignment,
-											onChange: function (alignment) {
-												props.setAttributes({alignment: alignment})
-											}
-										}
-									)
-									<?php }?>
-
-								),
-
-								el(wp.blockEditor.InspectorControls, {key: 'inspector'},
-
-									<?php
-
-									if(! empty( $this->arguments )){
-
-									if ( $show_advanced ) {
-									?>
-									el('div', {
-											style: {'padding-left': '16px','padding-right': '16px'}
-										},
-										el(
-											wp.components.ToggleControl,
-											{
-												label: 'Show Advanced Settings?',
-												checked: props.attributes.show_advanced,
-												onChange: function (show_advanced) {
-													props.setAttributes({show_advanced: !props.attributes.show_advanced})
-												}
-											}
-										)
-									)
-									,
-									<?php
-
-									}
-
-									$arguments = $this->group_arguments( $this->arguments );
-
-									// Do we have sections?
-									$has_sections = $arguments == $this->arguments ? false : true;
-
-
-									if($has_sections){
-									$panel_count = 0;
-									foreach($arguments as $key => $args){
-									?>
-									el(wp.components.PanelBody, {
-											title: '<?php esc_attr_e( $key ); ?>',
-											initialOpen: <?php if ( $panel_count ) {
-											echo "false";
-										} else {
-											echo "true";
-										}?>
-										},
-										<?php
-
-
-
-										foreach ( $args as $k => $a ) {
-
-											$this->block_row_start( $k, $a );
-											$this->build_block_arguments( $k, $a );
-											$this->block_row_end( $k, $a );
-										}
-										?>
-									),
-									<?php
-									$panel_count ++;
-
-									}
-									}else {
-									?>
-									el(wp.components.PanelBody, {
-											title: '<?php esc_attr_e( "Settings" ); ?>',
-											initialOpen: true
-										},
-										<?php
-										foreach ( $this->arguments as $key => $args ) {
-											$this->block_row_start( $key, $args );
-											$this->build_block_arguments( $key, $args );
-											$this->block_row_end( $key, $args );
-										}
-										?>
-									),
-									<?php
-									}
-
-									}
-									?>
-
-								),
-
-								<?php
-								// If the user sets block-output array then build it
-								if ( ! empty( $this->options['block-output'] ) ) {
-								$this->block_element( $this->options['block-output'] );
-							}else{
-								// if no block-output is set then we try and get the shortcode html output via ajax.
-								?>
-								el('div', {
-									dangerouslySetInnerHTML: {__html: onChangeContent()},
-									className: props.className,
-									style: {'minHeight': '30px'}
-								})
-								<?php
-								}
-								?>
-							]; // end return
-						},
-
-						// The "save" property must be specified and must be a valid function.
-						save: function (props) {
-
-							//console.log(props);
-
-
-							var attr = props.attributes;
-							var align = '';
-
-							// build the shortcode.
-							var content = "[<?php echo $this->options['base_id'];?>";
-							$html = '';
-							<?php
-
-							if(! empty( $this->arguments )){
-
-							foreach($this->arguments as $key => $args){
-							?>
-							if (attr.hasOwnProperty("<?php echo esc_attr( $key );?>")) {
-								if ('<?php echo esc_attr( $key );?>' == 'html') {
-									$html = attr.<?php echo esc_attr( $key );?>;
-								} else {
-									content += " <?php echo esc_attr( $key );?>='" + attr.<?php echo esc_attr( $key );?>+ "' ";
-								}
-							}
-							<?php
-							}
-							}
-
-							?>
-							content += "]";
-
-							// if has html element
-							if ($html) {
-								content += $html + "[/<?php echo $this->options['base_id'];?>]";
-							}
-
-
-							// @todo should we add inline style here or just css classes?
-							if (attr.alignment) {
-								if (attr.alignment == 'left') {
-									align = 'alignleft';
-								}
-								if (attr.alignment == 'center') {
-									align = 'aligncenter';
-								}
-								if (attr.alignment == 'right') {
-									align = 'alignright';
-								}
-							}
-
-							<?php
-							if(isset( $this->options['block-wrap'] ) && $this->options['block-wrap'] == ''){
-							?>
-							return content;
-							<?php
-							}else{
-							?>
-							var block_wrap = 'div';
-							if (attr.hasOwnProperty("block_wrap")) {
-								block_wrap = attr.block_wrap;
-							}
-							return el(block_wrap, {dangerouslySetInnerHTML: {__html: content}, className: align});
-							<?php
-							}
-							?>
-
-
-						}
-					});
-				})();
-			</script>
-			<?php
-			$output = ob_get_clean();
-
-			/*
-			 * We only add the <script> tags for code highlighting, so we strip them from the output.
-			 */
-
-			return str_replace( array(
-				'<script>',
-				'</script>'
-			), '', $output );
-		}
-
-		public function block_row_start($key, $args){
-
-			// check for row
-			if(!empty($args['row'])){
-
-				if(!empty($args['row']['open'])){
-
-				// element require
-				$element_require = ! empty( $args['element_require'] ) ? $this->block_props_replace( $args['element_require'], true ) . " && " : "";
-				echo $element_require;
-
-					if(false){?><script><?php }?>
-						el('div', {
-								className: 'bsui components-base-control',
-							},
-							<?php if(!empty($args['row']['title'])){ ?>
-							el('label', {
-									className: 'components-base-control__label',
-								},
-								'<?php echo addslashes( $args['row']['title'] ); ?>'
-							),
-							<?php }?>
-							<?php if(!empty($args['row']['desc'])){ ?>
-							el('p', {
-									className: 'components-base-control__help mb-0',
-								},
-								'<?php echo addslashes( $args['row']['desc'] ); ?>'
-							),
-							<?php }?>
-							el(
-								'div',
-								{
-									className: 'row mb-n2 <?php if(!empty($args['row']['class'])){ echo esc_attr($args['row']['class']);} ?>',
-								},
-								el(
-									'div',
-									{
-										className: 'col pr-2',
-									},
-
-					<?php
-					if(false){?></script><?php }
-				}elseif(!empty($args['row']['close'])){
-					if(false){?><script><?php }?>
-						el(
-							'div',
-							{
-								className: 'col pl-0',
-							},
-					<?php
-					if(false){?></script><?php }
-				}else{
-					if(false){?><script><?php }?>
-						el(
-							'div',
-							{
-								className: 'col pl-0 pr-2',
-							},
-					<?php
-					if(false){?></script><?php }
-				}
-
-			}
-
-		}
-
-		public function block_row_end($key, $args){
-
-			if(!empty($args['row'])){
-				// maybe close
-				if(!empty($args['row']['close'])){
-					echo "))";
-				}
-
-				echo "),";
-			}
-		}
-
-		public function build_block_arguments( $key, $args ) {
-			$custom_attributes = ! empty( $args['custom_attributes'] ) ? $this->array_to_attributes( $args['custom_attributes'] ) : '';
-			$options           = '';
-			$extra             = '';
-			$require           = '';
-
-			// `content` is a protected and special argument
-			if ( $key == 'content' ) {
-				return;
-			}
-
-
-			// icon
-			$icon = '';
-			if( !empty( $args['icon'] ) ){
-				$icon .= "el('div', {";
-									$icon .= "dangerouslySetInnerHTML: {__html: '".self::get_widget_icon( esc_attr($args['icon']))."'},";
-									$icon .= "className: 'text-center',";
-									$icon .= "title: '".addslashes( $args['title'] )."',";
-								$icon .= "}),";
-
-				// blank title as its added to the icon.
-				$args['title'] = '';
-			}
-
-			// require advanced
-			$require_advanced = ! empty( $args['advanced'] ) ? "props.attributes.show_advanced && " : "";
-
-			// element require
-			$element_require = ! empty( $args['element_require'] ) ? $this->block_props_replace( $args['element_require'], true ) . " && " : "";
-
-
-			$onchange  = "props.setAttributes({ $key: $key } )";
-			$onchangecomplete  = "";
-			$value     = "props.attributes.$key";
-			$text_type = array( 'text', 'password', 'number', 'email', 'tel', 'url', 'colorx' );
-			if ( in_array( $args['type'], $text_type ) ) {
-				$type = 'TextControl';
-				// Save numbers as numbers and not strings
-				if ( $args['type'] == 'number' ) {
-					$onchange = "props.setAttributes({ $key: Number($key) } )";
-				}
-			}
-			/*
-			 * https://www.wptricks.com/question/set-current-tab-on-a-gutenberg-tabpanel-component-from-outside-that-component/ es5 layout
-						elseif($args['type']=='tabs'){
-							?>
-							<script>
-								el(
-									wp.components.TabPanel,
-									{
-										tabs: [
-											{
-												name: 'show',
-												title: __( 'Show', 'my-textdomain' ),
-											},
-											{
-												name: 'edit',
-												title: __( 'Edit', 'my-textdomain' ),
-											},
-										],
-									},
-									( tab ) => {
-
-									if('show' === tab.name){
-									return 123;
-								}else if ( 'edit' === tab.name ) {
-									return 321;
-								}
-
-								}
-								),
-							</script>
-							<?php
-							return;
-						}
-			*/
-			elseif ( $args['type'] == 'color' ) {
-				$type = 'ColorPicker';
-				$onchange = "";
-				$extra = "color: $value,";
-				if(!empty($args['disable_alpha'])){
-					$extra .= "disableAlpha: true,";
-				}
-				$onchangecomplete = "onChangeComplete: function($key) {
-				value =  $key.rgb.a && $key.rgb.a < 1 ? \"rgba(\"+$key.rgb.r+\",\"+$key.rgb.g+\",\"+$key.rgb.b+\",\"+$key.rgb.a+\")\" : $key.hex;
-                        props.setAttributes({
-                            $key: value
-                        });
-                    },";
-			}
-			elseif ( $args['type'] == 'checkbox' ) {
-				$type = 'CheckboxControl';
-				$extra .= "checked: props.attributes.$key,";
-				$onchange = "props.setAttributes({ $key: ! props.attributes.$key } )";
-			} elseif ( $args['type'] == 'textarea' ) {
-				$type = 'TextareaControl';
-			} elseif ( $args['type'] == 'select' || $args['type'] == 'multiselect' ) {
-				$type = 'SelectControl';
-
-				if($args['name'] == 'category' && !empty($args['post_type_linked'])){
-					$options .= "options: taxonomies_".str_replace("-","_", $this->id).",";
-				}elseif($args['name'] == 'sort_by' && !empty($args['post_type_linked'])){
-					$options .= "options: sort_by_".str_replace("-","_", $this->id).",";
-				}else {
-
-					if ( ! empty( $args['options'] ) ) {
-						$options .= "options: [";
-						foreach ( $args['options'] as $option_val => $option_label ) {
-							$options .= "{ value: '" . esc_attr( $option_val ) . "', label: '" . addslashes( $option_label ) . "' },";
-						}
-						$options .= "],";
-					}
-				}
-				if ( isset( $args['multiple'] ) && $args['multiple'] ) { //@todo multiselect does not work at the moment: https://github.com/WordPress/gutenberg/issues/5550
-					$extra .= ' multiple: true, ';
-				}
-			} elseif ( $args['type'] == 'alignment' ) {
-				$type = 'AlignmentToolbar'; // @todo this does not seem to work but cant find a example
-			}elseif ( $args['type'] == 'margins' ) {
-
-			} else {
-				return;// if we have not implemented the control then don't break the JS.
-			}
-
-
-
-			// color input does not show the labels so we add them
-			if($args['type']=='color'){
-				// add show only if advanced
-				echo $require_advanced;
-				// add setting require if defined
-				echo $element_require;
-				echo "el('div', {style: {'marginBottom': '8px'}}, '".addslashes( $args['title'] )."'),";
-			}
-
-			// add show only if advanced
-			echo $require_advanced;
-			// add setting require if defined
-			echo $element_require;
-
-			// icon
-			echo $icon;
-			?>
-			el( wp.components.<?php echo $type; ?>, {
-			label: '<?php echo addslashes( $args['title'] ); ?>',
-			help: '<?php if ( isset( $args['desc'] ) ) {
-				echo addslashes( $args['desc'] );
-			} ?>',
-			value: <?php echo $value; ?>,
-			<?php if ( $type == 'TextControl' && $args['type'] != 'text' ) {
-				echo "type: '" . addslashes( $args['type'] ) . "',";
-			} ?>
-			<?php if ( ! empty( $args['placeholder'] ) ) {
-				echo "placeholder: '" . addslashes( $args['placeholder'] ) . "',";
-			} ?>
-			<?php echo $options; ?>
-			<?php echo $extra; ?>
-			<?php echo $custom_attributes; ?>
-			<?php echo $onchangecomplete;?>
-			onChange: function ( <?php echo $key; ?> ) {
-			<?php echo $onchange; ?>
-			}
-			} ),
-			<?php
-
-
-		}
-
 		/**
 		 * Convert an array of attributes to block string.
 		 *
@@ -2366,118 +677,6 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 			}
 
 			return $attributes;
-		}
-
-		/**
-		 * A self looping function to create the output for JS block elements.
-		 *
-		 * This is what is output in the WP Editor visual view.
-		 *
-		 * @param $args
-		 */
-		public function block_element( $args ) {
-
-
-			if ( ! empty( $args ) ) {
-				foreach ( $args as $element => $new_args ) {
-
-					if ( is_array( $new_args ) ) { // its an element
-
-
-						if ( isset( $new_args['element'] ) ) {
-
-							if ( isset( $new_args['element_require'] ) ) {
-								echo str_replace( array(
-										"'+",
-										"+'"
-									), '', $this->block_props_replace( $new_args['element_require'] ) ) . " &&  ";
-								unset( $new_args['element_require'] );
-							}
-
-							echo "\n el( '" . $new_args['element'] . "', {";
-
-							// get the attributes
-							foreach ( $new_args as $new_key => $new_value ) {
-
-
-								if ( $new_key == 'element' || $new_key == 'content' || $new_key == 'element_require' || $new_key == 'element_repeat' || is_array( $new_value ) ) {
-									// do nothing
-								} else {
-									echo $this->block_element( array( $new_key => $new_value ) );
-								}
-							}
-
-							echo "},";// end attributes
-
-							// get the content
-							$first_item = 0;
-							foreach ( $new_args as $new_key => $new_value ) {
-								if ( $new_key === 'content' || is_array( $new_value ) ) {
-
-									if ( $new_key === 'content' ) {
-										echo "'" . $this->block_props_replace( wp_slash( $new_value ) ) . "'";
-									}
-
-									if ( is_array( $new_value ) ) {
-
-										if ( isset( $new_value['element_require'] ) ) {
-											echo str_replace( array(
-													"'+",
-													"+'"
-												), '', $this->block_props_replace( $new_value['element_require'] ) ) . " &&  ";
-											unset( $new_value['element_require'] );
-										}
-
-										if ( isset( $new_value['element_repeat'] ) ) {
-											$x = 1;
-											while ( $x <= absint( $new_value['element_repeat'] ) ) {
-												$this->block_element( array( '' => $new_value ) );
-												$x ++;
-											}
-										} else {
-											$this->block_element( array( '' => $new_value ) );
-										}
-									}
-									$first_item ++;
-								}
-							}
-
-							echo ")";// end content
-
-							echo ", \n";
-
-						}
-					} else {
-
-						if ( substr( $element, 0, 3 ) === "if_" ) {
-							echo str_replace( "if_", "", $element ) . ": " . $this->block_props_replace( $new_args, true ) . ",";
-						} elseif ( $element == 'style' ) {
-							echo $element . ": " . $this->block_props_replace( $new_args ) . ",";
-						} else {
-							echo $element . ": '" . $this->block_props_replace( $new_args ) . "',";
-						}
-
-					}
-				}
-			}
-		}
-
-		/**
-		 * Replace block attributes placeholders with the proper naming.
-		 *
-		 * @param $string
-		 *
-		 * @return mixed
-		 */
-		public function block_props_replace( $string, $no_wrap = false ) {
-
-			if ( $no_wrap ) {
-				$string = str_replace( array( "[%", "%]" ), array( "props.attributes.", "" ), $string );
-			} else {
-				$string = str_replace( array( "[%", "%]" ), array( "'+props.attributes.", "+'" ), $string );
-			}
-
-			return $string;
 		}
 
 		/**
@@ -2689,6 +888,23 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 		}
 
 		/**
+		* Checks if the current call is a ajax call to get the block content.
+		*
+		* This can be used in your widget to return different content as the block content.
+		*
+		* @since 1.0.3
+		* @return bool
+		*/
+		public function is_block_content_call() {
+			$result = false;
+			if ( wp_doing_ajax() && isset( $_REQUEST['action'] ) && $_REQUEST['action'] == 'super_duper_output_shortcode' ) {
+				$result = true;
+			}
+
+			return $result;
+		}
+
+		/**
 		 * Output the super title.
 		 *
 		 * @param $args
@@ -2835,9 +1051,8 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 		 * @return string
 		 */
 		public function widget_advanced_toggle() {
-
 			$output = '';
-			if ( $this->block_show_advanced() ) {
+			if ( $this->block->block_show_advanced() ) {
 				$val = 1;
 			} else {
 				$val = 0;
@@ -3154,23 +1369,6 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 		}
 
 		/**
-		 * Checks if the current call is a ajax call to get the block content.
-		 *
-		 * This can be used in your widget to return different content as the block content.
-		 *
-		 * @since 1.0.3
-		 * @return bool
-		 */
-		public function is_block_content_call() {
-			$result = false;
-			if ( wp_doing_ajax() && isset( $_REQUEST['action'] ) && $_REQUEST['action'] == 'super_duper_output_shortcode' ) {
-				$result = true;
-			}
-
-			return $result;
-		}
-
-		/**
 		 * Get an instance hash that will be unique to the type and settings.
 		 *
 		 * @since 1.0.20
@@ -3202,9 +1400,7 @@ if ( ! class_exists( 'WP_Super_Duper' ) ) {
 				$css .= "</style>";
 			}
 
-
 			return $css;
-
 		}
 
 	}
